Re-create Login Attempts phinx migration.
Somehow the original login attempts migration did not get saved in git.  This commit recreates it, making some slight modificaitons along the way.  It now includes a foreign key to the user table; a null in this field indicates login failure.  It also includes a boolean to indicate whether the "remember me" box was checked, and a timestamp field to record explicit logout.  Not sure yet how to handle session logouts, if that's even possible.